+++ src/main/java/racingcar/domain/engine/Engine.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.engine;
 
 public interface Engine {
