+++ src/main/java/racingcar/domain/engine/powerstrategy/PowerStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.engine.powerstrategy;
 
 public interface PowerStrategy {
