+++ src/main/java/racingcar/domain/engine/powerstrategy/RandomPowerStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.engine.powerstrategy;
 
 import java.util.Random;
